+++ src/Control/Concurrent/NQE/PubSub.hs
@@ -1,10 +1,9 @@
 {-# LANGUAGE FlexibleContexts #-}
 module Control.Concurrent.NQE.PubSub
-    ( Publisher
-    , publisher
-    , boundedPublisher
+    ( publisher
+    , subscribe
+    , unsubscribe
     , withPubSub
-    , withBoundedPubSub
     ) where
 
 import           Control.Applicative
@@ -13,55 +12,41 @@ import           Control.Monad.Reader
 import           Data.List
 import           UnliftIO
 
-data ControlMsg ch msg
-    = Subscribe (ch msg)
-    | Unsubscribe (ch msg)
+-- | Subscribe or unsubscribe from an event publisher.
+data ControlMsg event
+    = Subscribe (Inbox event)
+    | Unsubscribe (Inbox event)
 
-data Incoming ch msg
-    = Control (ControlMsg ch msg)
-    | Event msg
+-- | Wrapper for a 'Mailbox' that can receive 'ControlMsg'.
+type Publisher msg = Inbox (ControlMsg msg)
 
-type Publisher mbox ch msg = mbox (ControlMsg ch msg)
-
+-- | Subscribe an 'Inbox' to events from a 'Publisher'. Unsubscribe when action
+-- finishes. Pass a mailbox constructor.
 withPubSub ::
-       (MonadUnliftIO m, Mailbox mbox)
-    => Publisher mbox TQueue msg
-    -> (TQueue msg -> m a)
+       (MonadUnliftIO m, Mailbox mbox event)
+    => Publisher event
+    -> m (mbox event)         -- ^ mailbox creator for this subscription
+    -> (Inbox event -> m a)   -- ^ unsubscribe when this action ends
     -> m a
-withPubSub pub f = bracket subscribe unsubscribe action
+withPubSub pub sub = bracket acquire (unsubscribe pub)
   where
-    subscribe = do
-        mbox <- newTQueueIO
-        Subscribe mbox `send` pub
-        return mbox
-    unsubscribe mbox = Unsubscribe mbox `send` pub
-    action mbox = f mbox
+    acquire = do
+        i <- newInbox =<< sub
+        subscribe pub i
+        return i
 
-withBoundedPubSub ::
-       (MonadUnliftIO m, Mailbox mbox)
-    => Int
-    -> Publisher mbox TBQueue msg
-    -> (TBQueue msg -> m a)
-    -> m a
-withBoundedPubSub bound pub f = bracket subscribe unsubscribe action
-  where
-    subscribe = do
-        mbox <- newTBQueueIO bound
-        Subscribe mbox `send` pub
-        return mbox
-    unsubscribe mbox = Unsubscribe mbox `send` pub
-    action mbox = f mbox
+-- | Subscribe an 'Inbox' to a 'Publisher' generating events.
+subscribe :: MonadIO m => Publisher event -> Inbox event -> m ()
+subscribe pub sub = Subscribe sub `send` pub
 
-publisher ::
-       ( MonadIO m
-       , Mailbox mbox
-       , Mailbox events
-       , Mailbox ch
-       , Eq (ch msg)
-       )
-    => Publisher mbox ch msg
-    -> events msg
-    -> m ()
+-- | Unsubscribe an 'Inbox' from a 'Publisher' events.
+unsubscribe :: MonadIO m => Publisher event -> Inbox event -> m ()
+unsubscribe pub sub = Unsubscribe sub `send` pub
+
+-- | Start a publisher that will receive events from an 'STM' action, and
+-- subscription requests from a 'Publisher' mailbox. Events will be forwarded
+-- atomically to all subscribers. Full mailboxes will not receive events.
+publisher :: MonadIO m => Publisher event -> STM event -> m ()
 publisher pub events = do
     box <- newTVarIO []
     runReaderT go box
@@ -70,62 +55,25 @@ publisher pub events = do
         forever $ do
             incoming <-
                 atomically $
-                Control <$> receiveSTM pub <|> Event <$> receiveSTM events
+                Left <$> receiveSTM pub <|> Right <$> events
             process incoming
 
-boundedPublisher ::
-       (MonadIO m, Mailbox mbox, Mailbox events)
-    => Publisher mbox TBQueue msg
-    -> events msg
-    -> m ()
-boundedPublisher pub events = do
-    box <- newTVarIO []
-    runReaderT go box
-  where
-    go =
-        forever $ do
-        incoming <-
-            atomically $
-            Control <$> receiveSTM pub <|> Event <$> receiveSTM events
-        processBound incoming
-
-processBound ::
-       (MonadIO m, MonadReader (TVar [TBQueue msg]) m)
-    => Incoming TBQueue msg
-    -> m ()
-processBound (Control (Subscribe mbox)) = do
-    box <- ask
-    atomically $ do
-        subscribers <- readTVar box
-        when (mbox `notElem` subscribers) $ writeTVar box (mbox : subscribers)
-
-processBound (Control (Unsubscribe mbox)) = do
-    box <- ask
-    atomically (modifyTVar box (delete mbox))
-
-processBound (Event event) =
-    ask >>= \box ->
-        atomically $
-        readTVar box >>= \subs ->
-            forM_ subs $ \sub ->
-                isFullTBQueue sub >>= \full ->
-                    when (not full) (event `sendSTM` sub)
-
 process ::
-       (Eq (ch msg), Mailbox ch, MonadIO m, MonadReader (TVar [ch msg]) m)
-    => Incoming ch msg
+       (MonadIO m, MonadReader (TVar [Inbox event]) m)
+    => Either (ControlMsg event) event
     -> m ()
-process (Control (Subscribe mbox)) = do
+process (Left (Subscribe sub)) = do
+    box <- ask
+    atomically (modifyTVar box (`union` [sub]))
+
+process (Left (Unsubscribe sub)) = do
+    box <- ask
+    atomically (modifyTVar box (delete sub))
+
+process (Right event) = do
     box <- ask
     atomically $ do
-        subscribers <- readTVar box
-        when (mbox `notElem` subscribers) $
-            writeTVar box (mbox : subscribers)
-
-process (Control (Unsubscribe mbox)) = do
-    box <- ask
-    atomically (modifyTVar box (delete mbox))
-
-process (Event event) = do
-    box <- ask
-    readTVarIO box >>= mapM_ (send event)
+        subs <- readTVar box
+        forM_ subs $ \sub -> do
+            full <- mailboxFullSTM sub
+            unless full $ event `sendSTM` sub