+++ src/NQE.hs
@@ -1,11 +1,11 @@
-module Control.Concurrent.NQE
+module NQE
     ( module Control.Concurrent.NQE.Process
-    , module Control.Concurrent.NQE.Network
     , module Control.Concurrent.NQE.Supervisor
     , module Control.Concurrent.NQE.PubSub
+    , module Control.Concurrent.NQE.Conduit
     ) where
 
-import           Control.Concurrent.NQE.Network
+import           Control.Concurrent.NQE.Conduit
 import           Control.Concurrent.NQE.Process
 import           Control.Concurrent.NQE.PubSub
 import           Control.Concurrent.NQE.Supervisor